[CWS] move accessors generator out of the secl package
