+++ pkg/security/generators/accessors/accessors.go
@@ -30,8 +30,8 @@ import (
 	"golang.org/x/text/language"
 	"golang.org/x/tools/go/packages"
 
-	"github.com/DataDog/datadog-agent/pkg/security/secl/compiler/generators/accessors/common"
-	"github.com/DataDog/datadog-agent/pkg/security/secl/compiler/generators/accessors/doc"
+	"github.com/DataDog/datadog-agent/pkg/security/generators/accessors/common"
+	"github.com/DataDog/datadog-agent/pkg/security/generators/accessors/doc"
 )
 
 const (
