+++ pkg/security/generators/accessors/common/types.go
