+++ pkg/security/generators/accessors/doc/doc.go
@@ -18,7 +18,7 @@ import (
 
 	"golang.org/x/tools/go/packages"
 
-	"github.com/DataDog/datadog-agent/pkg/security/secl/compiler/generators/accessors/common"
+	"github.com/DataDog/datadog-agent/pkg/security/generators/accessors/common"
 )
 
 const (
