+++ tasks/security_agent.py
@@ -541,9 +541,9 @@ def generate_cws_documentation(ctx, go_generate=False):
 def cws_go_generate(ctx, verbose=False):
     ctx.run("go install golang.org/x/tools/cmd/stringer")
     ctx.run("go install github.com/mailru/easyjson/easyjson")
+    ctx.run("go install github.com/DataDog/datadog-agent/pkg/security/generators/accessors")
     ctx.run("go install github.com/DataDog/datadog-agent/pkg/security/generators/operators")
     with ctx.cd("./pkg/security/secl"):
-        ctx.run("go install github.com/DataDog/datadog-agent/pkg/security/secl/compiler/generators/accessors")
         if sys.platform == "linux":
             ctx.run("GOOS=windows go generate ./...")
         # Disable cross generation from windows for now. Need to fix the stringer issue.
